+++ paddle_billing/Resources/Shared/Operations/__init__.py
@@ -1,5 +1,3 @@
-from paddle_billing.Resources.Shared.Operations.Create.CreateBillingDetails import CreateBillingDetails
-from paddle_billing.Resources.Shared.Operations.Update.UpdateBillingDetails import UpdateBillingDetails
 from paddle_billing.Resources.Shared.Operations.List.Comparator import Comparator
 from paddle_billing.Resources.Shared.Operations.List.DateComparison import DateComparison
 from paddle_billing.Resources.Shared.Operations.List.OrderBy import OrderBy